🔥 Refactor part 1 - `falcon_core`
- Remove connection traits
- Remove packet errors
- Remove "falcon_core_derive"
