+++ crates/core/Cargo.toml
@@ -6,8 +6,6 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-falcon_core_derive = { path = "../core_derive" }
-
 tracing = "0.1.36"
 thiserror = "1.0.32"
