+++ crates/core/src/error.rs
@@ -1,6 +1,4 @@
-use crate::network::packet::TaskScheduleError;
 use crate::world::blocks::ParseBlockError;
-use std::string::FromUtf8Error;
 use thiserror::Error;
 
 pub type Result<T> = std::result::Result<T, FalconCoreError>;
@@ -11,18 +9,6 @@ pub enum FalconCoreError {
     IoError(#[from] std::io::Error),
     #[error("The packet length was longer than 21 bits")]
     PacketTooLong,
-    #[error("PacketBuffer reached EOF")]
-    NoMoreBytes,
-    #[error("Variable i32 was longer than 5 bytes")]
-    VarI32TooLong,
-    #[error("Variable i64 was longer than 10 bytes")]
-    VarI64TooLong,
-    #[error("Invalid utf-8")]
-    BadString(#[from] FromUtf8Error),
-    #[error("String size was 0")]
-    StringSizeZero,
-    #[error("String was longer than expected: {1} > {0}")]
-    StringTooLong(i32, i32),
     #[error("Schematic version {0} is not supported")]
     InvalidSchematic(i32),
     #[error("Invalid schematic data version, should be {0} instead of {1}")]
@@ -33,6 +19,4 @@ pub enum FalconCoreError {
     MissingData,
     #[error("Error while reading block data")]
     ParseBlockError(#[from] ParseBlockError),
-    #[error("Error while executing packet logic")]
-    PacketHandleError(#[from] TaskScheduleError),
 }
