+++ crates/core/src/lib.rs
@@ -11,4 +11,3 @@ pub mod world;
 mod shutdown;
 
 pub use shutdown::ShutdownHandle;
-
