--- crates/core_derive/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "falcon_core_derive"
-version = "0.2.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[lib]
-proc-macro = true
-
-[dependencies]
-syn = { version = "1.0.99", features = ["full"] }
-quote = "1.0.21"
-proc-macro2 = "1.0.43"
